* master: (88 commits)
  Typeo: mt_rand instead of mt_srand
  Update release readme
  Version Bump
  PHProjekt 6.0.4
  PHProjekt 6.0.4
  Allow unicode letters in the filter query.
  Remove the lowercase transform from the sanitizer and put it into the filter class.
  Allow unicode letters in the filter query.
  Fix notification for send only if the user have an email.
  Fix js error for null value.
  Update Zend version.
  Fix translation class for use an array as parameter.
  Update Zend Framework to 1.10.7
  Update dojo version in the README file.
  Fix buttons layout.
  Fix search button layout.
  Align HTML and Tooltips buttons.
  Fix layout for selectBox.
  Fix js error when the array is null.
  Change deprecated attr() to new get()/set() functions.
  ...

+++ phprojekt/application/Default/Controllers/IndexController.php
@@ -589,7 +589,7 @@ class IndexController extends Zend_Controller_Action
 
         $model = $this->getModelObject()->find($id);
 
-        if ($model instanceof Phprojekt_Model_Interface) {
+        if ($model instanceof Phprojekt_ActiveRecord_Abstract) {
             $tmp = Default_Helpers_Delete::delete($model);
             if ($tmp === false) {
                 $message    = Phprojekt::getInstance()->translate(self::DELETE_FALSE_TEXT);
@@ -640,10 +640,12 @@ class IndexController extends Zend_Controller_Action
             $model    = $this->getModelObject();
             $idsArray = explode(",", $ids);
 
-            foreach ($idsArray as $id) {
-                $model->find((int) $id);
-                Default_Helpers_Delete::delete($model);
-                $showId[] = $id;
+            if ($model instanceof Phprojekt_ActiveRecord_Abstract) {
+                foreach ($idsArray as $id) {
+                    $model->find((int) $id);
+                    Default_Helpers_Delete::delete($model);
+                    $showId[] = $id;
+                }
             }
 
             $return = array('type'    => 'success',
@@ -734,7 +736,7 @@ class IndexController extends Zend_Controller_Action
     }
 
     /**
-     * Returns the front configurations from the configuration.ini (front.xxx),
+     * Returns the front configurations from the configuration.php (front.xxx),
      * and some others Core Settings.
      *
      * The return is an array like ('name' => varName, 'value' => varValue')
